solid progress on state machine
